+++ amplify/backend/types/amplify-dependent-resources-ref.d.ts
@@ -1,0 +1,36 @@
+export type AmplifyDependentResourcesAttributes = {
+    "auth": {
+        "universalratings5a980c73": {
+            "IdentityPoolId": "string",
+            "IdentityPoolName": "string",
+            "UserPoolId": "string",
+            "UserPoolArn": "string",
+            "UserPoolName": "string",
+            "AppClientIDWeb": "string",
+            "AppClientID": "string"
+        },
+        "userPoolGroups": {
+            "AdminGroupRole": "string"
+        }
+    },
+    "api": {
+        "universalratings": {
+            "GraphQLAPIIdOutput": "string",
+            "GraphQLAPIEndpointOutput": "string"
+        }
+    },
+    "function": {
+        "universalratings5a980c73PostConfirmation": {
+            "Name": "string",
+            "Arn": "string",
+            "LambdaExecutionRole": "string",
+            "Region": "string"
+        },
+        "ShowAddedDiscordWebhookPublisher": {
+            "Name": "string",
+            "Arn": "string",
+            "Region": "string",
+            "LambdaExecutionRole": "string"
+        }
+    }
+}