+++ src/model/ModalType.js
@@ -1,0 +1,4 @@
+export default class ModalType {
+  static PROFILE = 1;
+  static SETTINGS = 2;
+}